requirements.txt
appdirs==1.4.4
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
datetime-truncate==1.1.1
escape==1.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
logzero==1.7.0
MarkupSafe==3.0.2
multitasking==0.0.11
numpy==2.2.1
pandas==2.2.3
peewee==3.17.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
pywin32==308
requests==2.32.3
six==1.17.0
smartapi-python==1.5.2
SQLAlchemy==2.0.36
timeseries==0.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
zope.interface==6.1
